+++ src/app/auth/_components/Modal/AdminLoginModal.tsx
@@ -1,8 +1,8 @@
 'use client';
 
 import { FC, useRef } from 'react';
-import Button from '../Button';
-import Input from '../Input';
+import Button from '../../../../components/Button';
+import Input from '../../../../components/Input';
 import Modal from './Modal';
 
 type AdminLoginModalProps = {
