+++ src/app/auth/_components/Modal/Modal.tsx
@@ -1,6 +1,12 @@
 'use client';
 
-import { Dialog, Transition } from '@headlessui/react';
+import {
+  Dialog,
+  Transition,
+  TransitionChild,
+  DialogPanel,
+  DialogTitle
+} from '@headlessui/react';
 import React, { Fragment } from 'react';
 import { IoClose } from 'react-icons/io5';
 
@@ -13,9 +19,9 @@ interface ModalProps {
 
 const Modal: React.FC<ModalProps> = ({ isOpen, onClose, children, title }) => {
   return (
-    <Transition.Root show={isOpen} as={Fragment}>
+    <Transition show={isOpen} as={Fragment}>
       <Dialog as="div" className="relative z-50" onClose={onClose}>
-        <Transition.Child
+        <TransitionChild
           as={Fragment}
           enter="ease-out duration-300"
           enterFrom="opacity-0"
@@ -25,11 +31,11 @@ const Modal: React.FC<ModalProps> = ({ isOpen, onClose, children, title }) => {
           leaveTo="opacity-0"
         >
           <div className="fixed inset-0 bg-gray-500 bg-opacity-[0.75] transition-opacity" />
-        </Transition.Child>
+        </TransitionChild>
 
         <div className="fixed inset-0 z-10 overflow-y-auto">
           <div className="flex min-h-full items-center justify-center p-4 text-center sm:p-0">
-            <Transition.Child
+            <TransitionChild
               as={Fragment}
               enter="ease-out duration-300"
               enterFrom="opacity-0 translate-y-4 sm:translate-y-0 sm:scale-95"
@@ -38,8 +44,8 @@ const Modal: React.FC<ModalProps> = ({ isOpen, onClose, children, title }) => {
               leaveFrom="opacity-100 translate-y-0 sm:scale-100"
               leaveTo="opacity-0 translate-y-4 sm:translate-y-0 sm:scale-95"
             >
-              <Dialog.Panel className="relative w-full overflow-hidden rounded-lg bg-white px-4 pb-4 pt-5 text-left shadow-xl transition-all sm:my-8 sm:w-full sm:max-w-lg sm:p-6">
-                <Dialog.Title className={`text-2xl`}>{title}</Dialog.Title>
+              <DialogPanel className="relative w-full overflow-hidden rounded-lg bg-white px-4 pb-4 pt-5 text-left shadow-xl transition-all sm:my-8 sm:w-full sm:max-w-lg sm:p-6">
+                <DialogTitle className={`text-2xl`}>{title}</DialogTitle>
                 <div className="absolute right-0 top-0 z-10 hidden pr-4 pt-4 sm:block">
                   <button
                     type="button"
@@ -51,12 +57,12 @@ const Modal: React.FC<ModalProps> = ({ isOpen, onClose, children, title }) => {
                   </button>
                 </div>
                 {children}
-              </Dialog.Panel>
-            </Transition.Child>
+              </DialogPanel>
+            </TransitionChild>
           </div>
         </div>
       </Dialog>
-    </Transition.Root>
+    </Transition>
   );
 };
